requirements.txt
asgiref==3.4.1
Babel==2.10.3
certifi==2022.6.15
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.14
django-cors-headers==3.13.0
django-filter==21.1
django-money==3.0.0
django-parler==2.3
django-parler-rest==2.2
django-tenants==3.4.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
idna==3.3
importlib-metadata==4.8.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.7
MarkupSafe==2.0.1
packaging==21.3
psycopg2-binary==2.9.3
py-moneyed==2.0
PyJWT==2.4.0
pyparsing==3.0.7
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.10
zipp==3.6.0
